Rename licence application components
