+++ src/Spd.Manager.Licence.UnitTest/LicenceAppDocumentManagerTest.cs
@@ -4,7 +4,7 @@ using Microsoft.AspNetCore.Http;
 using Moq;
 using Spd.Resource.Repository.BizLicApplication;
 using Spd.Resource.Repository.Document;
-using Spd.Resource.Repository.LicenceApplication;
+using Spd.Resource.Repository.PersonLicApplication;
 using Spd.Utilities.TempFileStorage;
 
 namespace Spd.Manager.Licence.UnitTest;
