+++ src/Spd.Manager.Licence/ApplicantProfileManager.cs
@@ -6,7 +6,7 @@ using Spd.Resource.Repository.Application;
 using Spd.Resource.Repository.Contact;
 using Spd.Resource.Repository.Document;
 using Spd.Resource.Repository.Identity;
-using Spd.Resource.Repository.LicenceApplication;
+using Spd.Resource.Repository.PersonLicApplication;
 using Spd.Resource.Repository.Registration;
 using Spd.Utilities.Shared.Exceptions;
 using System.Net;
