+++ src/Spd.Manager.Licence/LicenceAppDocumentManager.cs
@@ -3,7 +3,7 @@ using MediatR;
 using Spd.Resource.Repository.Application;
 using Spd.Resource.Repository.BizLicApplication;
 using Spd.Resource.Repository.Document;
-using Spd.Resource.Repository.LicenceApplication;
+using Spd.Resource.Repository.PersonLicApplication;
 using Spd.Utilities.TempFileStorage;
 
 namespace Spd.Manager.Licence;
