+++ src/Spd.Manager.Printing/Documents/TransformationStrategies/Mappings.cs
@@ -2,7 +2,7 @@
 using Spd.Resource.Repository.Application;
 using Spd.Resource.Repository.Contact;
 using Spd.Resource.Repository.Licence;
-using Spd.Resource.Repository.LicenceApplication;
+using Spd.Resource.Repository.PersonLicApplication;
 using Spd.Resource.Repository.Org;
 
 namespace Spd.Manager.Printing.Documents.TransformationStrategies
