+++ src/Spd.Resource.Repository.IntegrationTest/BizLicApplicationRepositoryTest.cs
@@ -2,7 +2,7 @@
 using Microsoft.Dynamics.CRM;
 using Microsoft.Extensions.DependencyInjection;
 using Spd.Resource.Repository.BizLicApplication;
-using Spd.Resource.Repository.LicenceApplication;
+using Spd.Resource.Repository.PersonLicApplication;
 using Spd.Utilities.Dynamics;
 using Spd.Utilities.Shared.Exceptions;
