+++ src/Spd.Resource.Repository/ServiceExtensions.cs
@@ -15,7 +15,7 @@ using Spd.Resource.Repository.Identity;
 using Spd.Resource.Repository.Incident;
 using Spd.Resource.Repository.Invoice;
 using Spd.Resource.Repository.Licence;
-using Spd.Resource.Repository.LicenceApplication;
+using Spd.Resource.Repository.PersonLicApplication;
 using Spd.Resource.Repository.LicenceFee;
 using Spd.Resource.Repository.OptionSet;
 using Spd.Resource.Repository.Org;
