+++ src/Spd.Resource.Repository/WorkerLicenceCategory/Contract.cs
@@ -1,4 +1,4 @@
-using Spd.Resource.Repository.LicenceApplication;
+using Spd.Resource.Repository.PersonLicApplication;
 
 namespace Spd.Resource.Repository.WorkerLicenceCategory;
 public partial interface IWorkerLicenceCategoryRepository
