+++ src/Spd.Resource.Repository/WorkerLicenceCategory/Mappings.cs
@@ -1,6 +1,6 @@
 ﻿using AutoMapper;
 using Microsoft.Dynamics.CRM;
-using Spd.Resource.Repository.LicenceApplication;
+using Spd.Resource.Repository.PersonLicApplication;
 using Spd.Utilities.Dynamics;
 
 namespace Spd.Resource.Repository.WorkerLicenceCategory;
